autotag: move tests to test_hooks.py
